refactor for HOT-1543 Screening contact validation rule: No Allegations
